fix: cosf -> cos, sinf -> sin

+++ analysis/tests/simd/vec/AVX/avx_float_vec2_tests.cpp
@@ -295,7 +295,7 @@ namespace tests
 
         for (size_t i = 0; i < vecLen; i++)
         {
-            float expected = std::sinf(vecData[i]);
+            float expected = std::sin(vecData[i]);
             EXPECT_NEAR(res.data[i], expected, maximalSinCosAbsoluteError * std::abs(expected == 0 ? 0.1 : expected));
         }
     }
@@ -310,7 +310,7 @@ namespace tests
 
         for (size_t i = 0; i < vecLen; i++)
         {
-            float expected = std::sinf(vecData[i]);
+            float expected = std::sin(vecData[i]);
             EXPECT_NEAR(res.data[i], expected, maximalSinCosAbsoluteError * std::abs(expected == 0 ? 0.1 : expected));
         }
     }
@@ -325,7 +325,7 @@ namespace tests
 
         for (size_t i = 0; i < vecLen; i++)
         {
-            float expected = std::cosf(vecData[i]);
+            float expected = std::cos(vecData[i]);
             EXPECT_NEAR(res.data[i], expected, maximalSinCosAbsoluteError * std::abs(expected == 0 ? 0.1 : expected));
         }
     }
@@ -340,7 +340,7 @@ namespace tests
 
         for (size_t i = 0; i < vecLen; i++)
         {
-            float expected = std::cosf(vecData[i]);
+            float expected = std::cos(vecData[i]);
             EXPECT_NEAR(res.data[i], expected, maximalSinCosAbsoluteError * std::abs(expected == 0 ? 0.1 : expected));
         }
     }
