+++ analysis/tests/simd/vec/AVX/avx_float_vec_tests.cpp
@@ -310,7 +310,7 @@ namespace tests
 
         for (size_t i = 0; i < vecLen; i++)
         {
-            float expected = std::sinf(vecData[i]);
+            float expected = std::sin(vecData[i]);
             EXPECT_NEAR(res.data[i], expected, maximalSinCosAbsoluteError * std::abs(expected == 0 ? 0.1 : expected));
         }
     }
@@ -326,7 +326,7 @@ namespace tests
 
         for (size_t i = 0; i < vecLen; i++)
         {
-            float expected = std::sinf(vecData[i]);
+            float expected = std::sin(vecData[i]);
             EXPECT_NEAR(res.data[i], expected, maximalSinCosAbsoluteError * std::abs(expected == 0 ? 0.1 : expected));
         }
     }
@@ -341,7 +341,7 @@ namespace tests
 
         for (size_t i = 0; i < vecLen; i++)
         {
-            float expected = std::cosf(vecData[i]);
+            float expected = std::cos(vecData[i]);
             EXPECT_NEAR(res.data[i], expected, maximalSinCosAbsoluteError * std::abs(expected == 0 ? 0.1 : expected));
         }
     }
@@ -357,7 +357,7 @@ namespace tests
 
         for (size_t i = 0; i < vecLen; i++)
         {
-            float expected = std::cosf(vecData[i]);
+            float expected = std::cos(vecData[i]);
             EXPECT_NEAR(res.data[i], expected, maximalSinCosAbsoluteError * std::abs(expected == 0 ? 0.1 : expected));
         }
     }
