+++ analysis/tests/simd/vec/SEQ/seq_float_vec_tests.cpp
@@ -317,7 +317,7 @@ namespace tests
 
         for (size_t i = 0; i < vecLen; i++)
         {
-            EXPECT_FLOAT_EQ(res.data[i], std::sinf(vecData[i]));
+            EXPECT_FLOAT_EQ(res.data[i], std::sin(vecData[i]));
         }
     }
     TEST(FloatSeqVector, Vec9SinNegative)
@@ -332,7 +332,7 @@ namespace tests
 
         for (size_t i = 0; i < vecLen; i++)
         {
-            EXPECT_FLOAT_EQ(res.data[i], std::sinf(vecData[i]));
+            EXPECT_FLOAT_EQ(res.data[i], std::sin(vecData[i]));
         }
     }
     TEST(FloatSeqVector, Vec4CosPositive)
@@ -346,7 +346,7 @@ namespace tests
 
         for (size_t i = 0; i < vecLen; i++)
         {
-            EXPECT_FLOAT_EQ(res.data[i], std::cosf(vecData[i]));
+            EXPECT_FLOAT_EQ(res.data[i], std::cos(vecData[i]));
         }
     }
     TEST(FloatSeqVector, Vec9CosNegative)
@@ -361,7 +361,7 @@ namespace tests
 
         for (size_t i = 0; i < vecLen; i++)
         {
-            EXPECT_FLOAT_EQ(res.data[i], std::cosf(vecData[i]));
+            EXPECT_FLOAT_EQ(res.data[i], std::cos(vecData[i]));
         }
     }
     TEST(FloatSeqVector, Vec4TanPositive)
@@ -433,7 +433,7 @@ namespace tests
 
         for (size_t i = 0; i < vecLen; i++)
         {
-            float expected = std::sinf(vecData[i]);
+            float expected = std::sin(vecData[i]);
             EXPECT_NEAR(res.data[i], expected, maximalSinCosAbsoluteError * std::abs(expected == 0 ? 0.1 : expected));
         }
     }
@@ -449,7 +449,7 @@ namespace tests
 
         for (size_t i = 0; i < vecLen; i++)
         {
-            float expected = std::sinf(vecData[i]);
+            float expected = std::sin(vecData[i]);
             EXPECT_NEAR(res.data[i], expected, maximalSinCosAbsoluteError * std::abs(expected == 0 ? 0.1 : expected));
         }
     }
@@ -464,7 +464,7 @@ namespace tests
 
         for (size_t i = 0; i < vecLen; i++)
         {
-            float expected = std::cosf(vecData[i]);
+            float expected = std::cos(vecData[i]);
             EXPECT_NEAR(res.data[i], expected, maximalSinCosAbsoluteError * std::abs(expected == 0 ? 0.1 : expected));
         }
     }
@@ -480,7 +480,7 @@ namespace tests
 
         for (size_t i = 0; i < vecLen; i++)
         {
-            float expected = std::cosf(vecData[i]);
+            float expected = std::cos(vecData[i]);
             EXPECT_NEAR(res.data[i], expected, maximalSinCosAbsoluteError * std::abs(expected == 0 ? 0.1 : expected));
         }
     }
